feat: changes the the way db perms are handled
This commit changes the way that the database permissions are handled.
Before cloning the database, the app now stores the existing user, group
and file permissions in the config.ini file. When it copies the backup
to the RAM disk, it reapplies these properties to the file, effectively
restoring its original state.

SECURITY NOTE
This update improves security by decreasing the attack the app's
potential attack surface by doing the following:
* The "docker" user no longer uses the "docker" password. Instead, a
  random secure password is set for it when the Docker image is built.
* The "docker" user effectively retains sudo via the passwordless method
  for only one command: chown. Since no one can know what the password
  is for the "docker" user, this blocks elevated access to other
  commands.

BREAKING CHANGE
The USE_ROOT environment variable was removed in favor of the much more
secure implementation detailed above.